Fix detection of pkg-config on Msys2
Replace call to _cmd_exists with shutil.where. This properly handles the
".exe" extension on windows build systems. There are no other uses of
_cmd_exists so it it removed.

This change fixes builds of pillow in python virtual enviroments.  When
running pip in a virtual env, sys.prefix will resolve to the root of the
virtual env and path searching for zlib and friends will fail.  The
usual work around is to force paths with env variables INCLUDE, and LIB.
Forcing pkg-config also works: PKG_CONFIG="pkg-config.exe".

With proper detection of pkg-config, this works as expected:
   pip install .